% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMgr.R
\name{modelMgr}
\alias{modelMgr}
\title{Graphical UI For The AMModels Package}
\usage{
modelMgr(...)
}
\arguments{
\item{\dots}{Additional arguments to \code{shiny::runApp}.}
}
\description{
The model manager is a graphical user interface supplied to accomplish nearly all functions within the AMModels package. Models and data can be tagged with metadata and organized within model libraries. Model libraries can be imported, searched, subset, edited, and exported through the GUI. The modelMgr GUI allows users to import models and data from either a .RData or .rda file or from the user's \code{.GlobalEnv}.
}
\details{
modelMgr
}
\examples{
\dontrun{ 
# The shiny app
modelMgr()

# Accepts args to shiny::runApp
modelMgr(quiet = TRUE)
}
}
\author{
Jon Katz
}
\keyword{misc}
