% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMModels-package.R
\docType{data}
\name{mymodels}
\alias{mymodels}
\title{Sample \code{amModelLib} Object Containing Models and Data.}
\format{An \code{amModelLib} object with}
\description{
Four models and four data objects.
}
\examples{

data(mymodels)
mymodels
summary(mymodels)
getAMModel(mymodels, 'plant.model')
mymodels[[1]]
str(mymodels)



}
\keyword{datasets}
